Add Exp. Candy and its functionalities

+++ include/item.h
@@ -11,7 +11,7 @@ struct Item
     u16 itemId;
     u16 price;
     u8 holdEffect;
-    u8 holdEffectParam;
+    u16 holdEffectParam;
     const u8 *description;
     u8 importance;
     u8 registrability;
